+++ pkg/river/internal/rivertags/rivertags.go
@@ -18,6 +18,7 @@ const (
 
 	FlagOptional // FlagOptional marks a field optional for decoding/encoding
 	FlagLabel    // FlagLabel will store block labels in the field
+	FlagSquash   // FlagSquash will expose inner fields from a struct as outer fields.
 )
 
 // String returns the flags as a string.
@@ -35,6 +36,9 @@ func (f Flags) String() string {
 	}
 	if f&FlagLabel != 0 {
 		attrs = append(attrs, "label")
+	}
+	if f&FlagSquash != 0 {
+		attrs = append(attrs, "squash")
 	}
 
 	return fmt.Sprintf("Flags(%s)", strings.Join(attrs, ","))
@@ -89,8 +93,11 @@ func (f Field) IsOptional() bool { return f.Flags&FlagOptional != 0 }
 //	// represents.
 //	Field string `river:",label"`
 //
-// With the exception of the `river:",label"` tag, all tagged fields must have a
-// unique name.
+//	// Attributes and blocks inside of Field are exposed as top-level fields.
+//	Field struct{} `river:",squash"`
+//
+// With the exception of the `river:",label"` and `river:",squash" tags, all
+// tagged fields must have a unique name.
 //
 // The type of tagged fields may be any Go type, with the exception of
 // `river:",label"` tags, which must be strings.
@@ -141,20 +148,11 @@ func Get(ty reflect.Type) []Field {
 		}
 		usedNames[fullName] = tf.Index
 
-		switch options[1] {
-		case "attr":
-			tf.Flags |= FlagAttr
-		case "attr,optional":
-			tf.Flags |= FlagAttr | FlagOptional
-		case "block":
-			tf.Flags |= FlagBlock
-		case "block,optional":
-			tf.Flags |= FlagBlock | FlagOptional
-		case "label":
-			tf.Flags |= FlagLabel
-		default:
+		flags, ok := parseFlags(options[1])
+		if !ok {
 			panic(fmt.Sprintf("river: unrecognized river tag format %q at %s", tag, printPathToField(ty, tf.Index)))
 		}
+		tf.Flags = flags
 
 		if len(tf.Name) > 1 && tf.Flags&FlagBlock == 0 {
 			panic(fmt.Sprintf("river: field names with `.` may only be used by blocks (found at %s)", printPathToField(ty, tf.Index)))
@@ -174,7 +172,27 @@ func Get(ty reflect.Type) []Field {
 			usedLabelField = tf.Index
 		}
 
-		if fullName == "" && tf.Flags&FlagLabel == 0 /* (e.g., *not* a label) */ {
+		if tf.Flags&FlagSquash != 0 {
+			if fullName != "" {
+				panic(fmt.Sprintf("river: squash field at %s must not have a name", printPathToField(ty, tf.Index)))
+			}
+
+			// Get the inner fields from the squashed struct and append each of them.
+			// The index of the squashed field is prepended to the index of the inner
+			// struct.
+			innerFields := Get(deferenceType(field.Type))
+			for _, innerField := range innerFields {
+				fields = append(fields, Field{
+					Name:  innerField.Name,
+					Index: append(field.Index, innerField.Index...),
+					Flags: innerField.Flags,
+				})
+			}
+
+			continue
+		}
+
+		if fullName == "" && tf.Flags&(FlagLabel|FlagSquash) == 0 /* (e.g., *not* a label or squash) */ {
 			panic(fmt.Sprintf("river: non-empty field name required at %s", printPathToField(ty, tf.Index)))
 		}
 
@@ -182,6 +200,27 @@ func Get(ty reflect.Type) []Field {
 	}
 
 	return fields
+}
+
+func parseFlags(input string) (f Flags, ok bool) {
+	switch input {
+	case "attr":
+		f |= FlagAttr
+	case "attr,optional":
+		f |= FlagAttr | FlagOptional
+	case "block":
+		f |= FlagBlock
+	case "block,optional":
+		f |= FlagBlock | FlagOptional
+	case "label":
+		f |= FlagLabel
+	case "squash":
+		f |= FlagSquash
+	default:
+		return
+	}
+
+	return f, true
 }
 
 func printPathToField(structTy reflect.Type, path []int) string {
@@ -202,4 +241,11 @@ func printPathToField(structTy reflect.Type, path []int) string {
 	}
 
 	return sb.String()
+}
+
+func deferenceType(ty reflect.Type) reflect.Type {
+	for ty.Kind() == reflect.Pointer {
+		ty = ty.Elem()
+	}
+	return ty
 }